requirements.txt
streamlit 
streamlit-lottie
ollama 
pyyaml 
reportlab